requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.5
